[description_display] revert to original display
"What's all this about?"

In the 'health' and 'adminimal medela' themes, the description field is invariably
output beneath its element.
We _were_ going to move descriptions for the various form elements to above
the element & beneath the label but the latest decision is to abandon this idea.

Thus, here I revert the edits introduced in <a5e2748f2c99> of 25 March 2021